requirements.txt
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
torchscale
timm
underthesea
salesforce-lavis
h5py
tqdm
Pillow
torch
transformers
mlflow
efficientnet_pytorch